+++ src/Attachment.php
@@ -302,21 +302,25 @@ class Attachment {
      * @return string|null
      */
     public function getExtension(): ?string {
+        $extension = null;
         $guesser = "\Symfony\Component\Mime\MimeTypes";
         if (class_exists($guesser) !== false) {
             /** @var Symfony\Component\Mime\MimeTypes $guesser */
             $extensions = $guesser::getDefault()->getExtensions($this->getMimeType());
-            return $extensions[0] ?? null;
+            $extension = $extensions[0] ?? null;
         }
-
-        $deprecated_guesser = "\Symfony\Component\HttpFoundation\File\MimeType\ExtensionGuesser";
-        if (class_exists($deprecated_guesser) !== false){
-            /** @var \Symfony\Component\HttpFoundation\File\MimeType\ExtensionGuesser $deprecated_guesser */
-            return $deprecated_guesser::getInstance()->guess($this->getMimeType());
+        if ($extension === null) {
+            $deprecated_guesser = "\Symfony\Component\HttpFoundation\File\MimeType\ExtensionGuesser";
+            if (class_exists($deprecated_guesser) !== false){
+                /** @var \Symfony\Component\HttpFoundation\File\MimeType\ExtensionGuesser $deprecated_guesser */
+                $extension = $deprecated_guesser::getInstance()->guess($this->getMimeType());
+            }
         }
-
-        $extensions = explode(".", $this->part->filename ?: $this->part->name);
-        return end($extensions);
+        if ($extension === null) {
+            $extensions = explode(".", $this->filename);
+            $extension = end($extensions);
+        }
+        return $extension;
     }
 
     /**